node's now have a score
only updates for asocial decisions in round 1. Note that str() and int() are used as properties start off a strings and need to be converted to ints before we can add them.

I'm not sure what your code was doing, but some issues:
"node.property1[-1]" I had no idea what this was doing at first, but now I see you are taking the last letter of their name, e.g. turning "A1" into just "1" and then comparing it to the info's origin. But remember that player A1 might not always be node 1. Also, you need to convert the string "1" into an int before doing the comparison, otherwise it always returns FALSE. So:

int(node.property1[-1]) == info.origin_id

would probably have worked. That said, this function only runs when a node has created an info and so this check is unnecessary: it will always return TRUE