first step success
Added n_copies per topic to LottyNode property1, and asoc_score per topic to LottyNode property1. Added topic to property1 of LottyInfo. Edited relevant parts in experiment.py and exp_shared.js and these properties now appear in the database in the right places and the experiment runs as usual. Next step is to increment score according to topic.