testing round2 warning div
this version takes you to the round 2 warning div on question 3 (question number needs changing in both round1.js and round2.js, to first proceed to round2.js, and then to display the round warning). This works but need to change the button contents, and make different for A,B and C. (Currently A&B were the same, and only C different).